pyproject.toml
[project]
name = "taichi-splatting"  
version = "0.30.0"  
description = "A Taichi Gaussian Splatting library"  
readme = "README.md" 
requires-python = ">=3.10"
license = {file = "LICENSE"}

maintainers = [
  {name = "Oliver Batchelor", email = "oliver.batchelor@canterbury.ac.nz" } 
]

dependencies = [
  "beartype", 
  "taichi-nightly", 
  "tqdm", 
  "tensordict"
]


[tool.setuptools.packages.find]
include = ["taichi_splatting"]
exclude = ["profiles", "benchmarks"]


[project.urls] 
"Homepage" = "https://github.com/uc-vision/taichi-splatting"

[build-system]
requires = ["setuptools>=43.0.0",  "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"


[options]
dependency_links = ["https://pypi.taichi.graphics/simple/"]

[project.scripts]  # Optional
fit_image_gaussians = "taichi_splatting.examples.fit_image_gaussians:main"
bench_projection = "taichi_splatting.benchmarks.bench_projection:main"
bench_rasterizer = "taichi_splatting.benchmarks.bench_rasterizer:main"
bench_tilemapper = "taichi_splatting.benchmarks.bench_tilemapper:main"

bench_sh = "taichi_splatting.benchmarks.bench_sh:main"


[tool.setuptools.package-data]

[tool.pyright]
reportInvalidTypeForm = "none"

[tool.ruff]
indent-width = 2

[tool.pytest.ini_options]
filterwarnings = [
    # disable "UserWarning: The .grad attribute of a Tensor that is not a leaf Tensor"
    "ignore::UserWarning"  
]

